chore(core-p2p): pet tslint, tweak log message
Also handle/allow verification of a chain with just 1 block, used by the
tests.
